# public/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "S5ARC. API Documentation",
    "description": "API documentation for S5ARC.",
    "version": "1.0.0",
    "contact": {
      "email": "s5sajidyt+s5arc_dev@gmail.com",
      "name": "S5SAJID (SajidUllah K.)",
      "url": "https://s5sajid.github.io"
    }
  },
  "paths": {
...
    "/api/v1/products/": {
      "get": {
        "operationId": "listProducts",
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProducts200ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProducts401ResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProducts403ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProducts404ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "Response for status 500",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProducts500ResponseBody"
                }
              }
            }
          }
        },
        "tags": ["Products"]
      },
      "post": {
        "operationId": "createProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Response for status 201",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProduct201ResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "Response for status 400",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProduct400ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProduct401ResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProduct403ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProduct404ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "Response for status 500",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProduct500ResponseBody"
                }
              }
            }
          }
        },
        "tags": ["Products"]
      }
    },
...
    "/api/v1/settings/": {
      "get": {
        "operationId": "getSettings",
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSettings200ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSettings401ResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSettings403ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSettings404ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "Response for status 500",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSettings500ResponseBody"
                }
              }
            }
          }
        },
        "tags": ["Settings"]
      },
      "patch": {
        "operationId": "updateSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSettings200ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSettings401ResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSettings403ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSettings404ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "Response for status 500",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSettings500ResponseBody"
                }
              }
            }
          }
        },
        "tags": ["Settings"]
      }
    }
  },
  "components": {
    "schemas": {
      "CreateProduct201ResponseBody": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "id": { "type": "string" }
        },
        "required": ["message"],
        "additionalProperties": false
      },
      "CreateProduct400ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "CreateProduct401ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "CreateProduct403ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "CreateProduct404ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "CreateProduct500ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "CreateProductRequestBody": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 2, "maxLength": 100 },
          "slug": { "type": "string", "minLength": 2, "maxLength": 100 },
          "description": {
            "type": "string",
            "minLength": 10,
            "maxLength": 1000
          },
          "price": { "type": "number", "exclusiveMinimum": true },
          "compare_at": { "type": "number", "exclusiveMinimum": true },
          "stock": { "type": "number" },
          "sku": { "type": "string" },
          "images": {
            "minItems": 1,
            "maxItems": 6,
            "type": "array",
            "items": {}
          },
          "is_published": { "default": false, "type": "boolean" }
        },
        "required": ["name", "slug", "description", "price", "images"]
      },
...
      "GetSettings200ResponseBody": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "name": { "type": "string", "maxLength": 255 },
              "domain": { "type": "string", "maxLength": 255 },
              "settings": {
                "anyOf": [
                  {
                    "anyOf": [
                      { "type": "string" },
                      { "type": "number" },
                      { "type": "boolean" },
                      { "type": "string", "nullable": true, "enum": [null] }
                    ]
                  },
                  { "type": "object", "additionalProperties": {} },
                  { "type": "array", "items": {} }
                ]
              }
            },
            "required": ["name", "domain", "settings"],
            "additionalProperties": false
          }
        },
        "required": ["data"],
        "additionalProperties": false
      },
      "GetSettings401ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "GetSettings403ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "GetSettings404ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "GetSettings500ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
...
      "ListProducts200ResponseBody": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "name": { "type": "string", "maxLength": 255 },
                "slug": { "type": "string", "maxLength": 255 },
                "description": { "type": "string" },
                "price": {
                  "type": "number",
                  "minimum": -8388608,
                  "maximum": 8388607
                },
                "compare_at": {
                  "nullable": true,
                  "type": "number",
                  "minimum": -8388608,
                  "maximum": 8388607
                },
                "stock": {
                  "nullable": true,
                  "type": "number",
                  "minimum": -8388608,
                  "maximum": 8388607
                },
                "sku": { "nullable": true, "type": "string", "maxLength": 255 },
                "images": {
                  "type": "array",
                  "items": { "type": "string", "maxLength": 300 }
                },
                "is_published": { "type": "boolean" },
                "created_at": {},
                "updated_at": {}
              },
              "required": [
                "id",
                "name",
                "slug",
                "description",
                "price",
                "compare_at",
                "stock",
                "sku",
                "images",
                "is_published",
                "created_at",
                "updated_at"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": ["data"],
        "additionalProperties": false
      },
      "ListProducts401ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "ListProducts403ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "ListProducts404ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "ListProducts500ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
...
      "UpdateSettings200ResponseBody": {
        "type": "object",
        "properties": { "message": { "type": "string" } },
        "required": ["message"],
        "additionalProperties": false
      },
      "UpdateSettings401ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "UpdateSettings403ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "UpdateSettings404ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "UpdateSettings500ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "UpdateSettingsRequestBody": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "maxLength": 255 },
          "domain": { "type": "string", "maxLength": 255 },
          "settings": {
            "anyOf": [
              {
                "anyOf": [
                  { "type": "string" },
                  { "type": "number" },
                  { "type": "boolean" },
                  { "type": "string", "nullable": true, "enum": [null] }
                ]
              },
              { "type": "object", "additionalProperties": {} },
              { "type": "array", "items": {} }
            ]
          }
        }
      }
    },
    "securitySchemes": {
      "x-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key required for authentication."
      }
    }
  },
...
  "security": [{ "x-api-key": [] }]
}